refact(modularization/inbounds): categorizing inbounds as module
